Move non-@Test to super, increase visibility (no other changes)
